move trace functions into seperate file, restore trace dump for 'too deep recursion' error.
